refactor: move review guidance contracts to output crate
